# backend/vibey-hub-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vibey Hub API",
    "version": "1.0.0",
    "description": "Internal contribution tracking + project management API for Vibey World."
  },
  "servers": [{ "url": "http://localhost:4000", "description": "Local dev" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "Supabase JWT" }
    },
    "schemas": {
...
      "Task": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "project_id": { "type": "string", "nullable": true },
          "assigned_user": { "type": "string", "nullable": true },
          "title": { "type": "string" },
          "priority": { "type": "string", "enum": ["low", "medium", "high"] },
          "status": { "type": "string", "enum": ["todo", "in_progress", "done"] },
          "points": { "type": "integer", "nullable": true },
          "deadline": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Contribution": {
        "type": "object",
        "properties": {
          "id": { "type": "string", "format": "uuid" },
          "user_id": { "type": "string" },
          "title": { "type": "string" },
          "category": { "type": "string" },
          "description": { "type": "string", "nullable": true },
          "points": { "type": "integer" },
          "evidence_url": { "type": "string", "nullable": true },
          "status": { "type": "string", "enum": ["pending", "approved", "rejected"] },
          "reviewer_note": { "type": "string", "nullable": true },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
...
    }
  },
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/tasks": {
      "get": {
        "tags": ["Tasks"],
        "parameters": [{ "name": "assigned_user", "in": "query", "schema": { "type": "string" } }],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Task" } } } } } }
      },
      "post": {
        "tags": ["Tasks"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title"],
                "properties": {
                  "title": { "type": "string" },
                  "assigned_user": { "type": "string" },
                  "priority": { "type": "string" },
                  "deadline": { "type": "string" },
                  "points": { "type": "integer" },
                  "project_id": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } } }
      }
    },
    "/api/tasks/{id}": {
      "patch": {
        "tags": ["Tasks"],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "status": { "type": "string" }, "priority": { "type": "string" }, "deadline": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Task" } } } } }
      }
    },
    "/api/contributions": {
      "get": {
        "tags": ["Contributions"],
        "parameters": [
          { "name": "user_id", "in": "query", "schema": { "type": "string" } },
          { "name": "status", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Contribution" } } } } } }
      },
      "post": {
        "tags": ["Contributions"],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["user_id", "title", "category", "difficulty"],
                "properties": {
                  "user_id": { "type": "string" },
                  "title": { "type": "string" },
                  "category": { "type": "string" },
                  "description": { "type": "string" },
                  "evidence_url": { "type": "string" },
                  "difficulty": { "type": "string", "enum": ["easy", "medium", "hard", "exceptional"] }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "Created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Contribution" } } } } }
      }
    },
    "/api/contributions/{id}": {
      "patch": {
        "tags": ["Contributions"],
        "summary": "Founder approves/rejects",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": { "status": { "type": "string", "enum": ["approved", "rejected", "pending"] }, "reviewer_note": { "type": "string" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "OK", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Contribution" } } } } }
      }
    },
...
  }
}